requirements.txt
agate==1.9.1
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
babel==2.17.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
daff==1.3.46
dbt-adapters==1.14.3
dbt-common==1.16.0
dbt-core==1.9.3
dbt-extractor==0.5.1
dbt-postgres==1.9.0
dbt-semantic-interfaces==0.7.4
deepdiff==7.0.1
et-xmlfile==2.0.0
gitdb==4.0.12
gitpython==3.1.44
greenlet==3.1.1
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
leather==0.4.0
markupsafe==3.0.2
mashumaro==3.14
more-itertools==10.6.0
msgpack==1.1.0
narwhals==1.31.0
networkx==3.4.2
numpy==2.2.4
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
parsedatetime==2.6
pathspec==0.12.1
pillow==11.1.0
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==19.0.1
pydantic==2.10.6
pydantic-core==2.27.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.1
pyyaml==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
six==1.17.0
smmap==5.0.2
snowplow-tracker==1.1.0
sqlalchemy==2.0.39
sqlparse==0.5.3
streamlit>=1.43.2
tenacity==9.0.0
text-unidecode==1.3
toml==0.10.2
tornado==6.4.2
typing-extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
zipp==3.21.0
plotly>=5.18.0
python-dotenv>=1.0.0
